Divide electronics between molecules and solid state
